Replace double tilde in README
GitHub doesn't render an org-mode expression that uses a double tilde, e.g.,
`~~/.local/share/yadm/repo.git~` - note the second tilde, which denotes the home
directory. This commit replaces the second tilde with `$HOME`. The yadm
documentation also uses `$HOME`.

This commit also adds a footnote about the location of the bare repo, which
changed from version 1 of yadm to version 2 and again from 2 to 3.